python-flask-server/requirements.txt
connexion == 1.1.15
python_dateutil == 2.6.0
setuptools >= 21.0.0
Werkzeug==0.16.1
openstacksdk
python-openstackclient